+++ models.json
@@ -7,7 +7,7 @@
             "optional_start": "(c3) <sp>V/</sp> O(e) Se(f)nhor(g) es(g)te(g)ja(e) con(f)vos(gf)co.(f) (::) <sp>R/</sp> E(e)l{e‿e}s(f)tá(g) no(g) mei(g)o(e) de(f) nós(gf) (::Z) <sp>V/</sp> Co(f)ra(g)ções(h) ao(g) al(fg)to.(fe) (::) <sp>R/</sp> O(g) nos(g)so(g) co(f)ra(g)ção(h) es(g)tá(f) em(g) Deus.(fe) (::Z) <sp>V/</sp> De(gf)mos(e) gra(ef)ças(g) ao(f) Se(g)nhor(hg) nos(fe)so(fg) Deus.(fgf) (::) <sp>R/</sp> É(f) nos(f)so(f) de(g)ver(h) e(g) nos(g)sa(f) sal(g)va(f)ção.(fe) (::Z) ",
             "end": "",
             "start": "(c3) <sp>V/</sp> ",
-            "default": "(gr gr gr) (fe) (ef) (g) (fgr1) (fr) (f) (:)",
+            "default": "(g) (gr gr gr) (fe) (ef) (g) (fgr1) (fr) (f) (:)",
             "patterns": [
                 {
                     "symbol": "*",
@@ -33,7 +33,7 @@
             "optional_start": "(c3) <sp>V/</sp> O(f) Se(g)nhor(h) es(h)te(h)ja(f) con(g)vos(h)co.(g) (::) <sp>R/</sp> E(f)l{e‿e}s(g)tá(h) no(h) mei(h)o(f) de(g) nós(hg) (::Z) <sp>V/</sp> Co(g)ra(h)ções(i) ao(h) al(g)to.(g) (::) <sp>R/</sp> O(h) nos(h)so(h) co(g)ra(h)ção(i) es(h)tá(g) em(h) Deus.(gf) (::Z) <sp>V/</sp> De(ih)mos(g) gra(h)ças(h) ao(h) Se(h)nhor(h) nos(f)so(g) Deus.(hg) (::) <sp>R/</sp> É(g) nos( g)so(g) de(h)ver(i) e(h) nos(h)sa(g) sal(h)va(g)ção.(gf) (::Z)",
             "end": "",
             "start": "(c3) <sp>V/</sp> ",
-            "default": "(gr gr gr) (e) (f) (gr1) (fr) (f) (:)",
+            "default": "(g) (gr gr gr) (e) (f) (gr1) (fr) (f) (:)",
             "patterns": [
                 {
                     "symbol": "*",
@@ -41,7 +41,7 @@
                 },
                 {
                     "symbol": "+",
-                    "gabc": "(gr gr gr) (hr1) (gr) (g) (,)"
+                    "gabc": "(g) (gr gr gr) (hr1) (gr) (g) (,)"
                 }
             ],
             "find": [],
@@ -89,15 +89,15 @@
             "optional_start": "(c3) <sp>V/</sp> O(h) re(h) mos:(f) (::)",
             "end": "",
             "start": "(c3) <sp>V/</sp> ",
-            "default": "(hr hr hr) (hr1) (dr) (d) (:)",
+            "default": "(h) (hr hr hr) (hr1) (dr) (d) (:)",
             "patterns": [
                 {
                     "symbol": "*",
-                    "gabc": "(hr hr hr) (g) (f) (hr1) (hr) (h) (;)"
+                    "gabc": "(h) (hr hr hr) (g) (f) (hr1) (hr) (h) (;)"
                 },
                 {
                     "symbol": "+",
-                    "gabc": "(hr hr hr) (hr1) (fr) (f) (,)"
+                    "gabc": "(h) (hr hr hr) (hr1) (fr) (f) (,)"
                 }
             ],
             "find": [
@@ -127,15 +127,15 @@
             "optional_start": "(c3) <sp>V/</sp> O(h) re(h) mos:(h) (::)",
             "end": "",
             "start": "(c3) <sp>V/</sp> ",
-            "default": "(hr hr hr) (hr1) (hr) (h) (:)",
+            "default": "(h) (hr hr hr) (hr1) (hr) (h) (:)",
             "patterns": [
                 {
                     "symbol": "*",
-                    "gabc": "(hr hr hr) (g) (f) (hr1) (hr) (h) (;)"
+                    "gabc": "(h) (hr hr hr) (g) (f) (hr1) (hr) (h) (;)"
                 },
                 {
                     "symbol": "+",
-                    "gabc": "(hr hr hr) (hr1) (gr) (g) (,)"
+                    "gabc": "(h) (hr hr hr) (hr1) (gr) (g) (,)"
                 }
             ],
             "find": [
@@ -165,11 +165,11 @@
             "optional_start": "O(h) Se(h)nhor(h) es(h)te(h)ja(h) con(h)vos(h)co.(h) (::Z)E(h)le(h) es(h)tá(h) no(h) meio(h) de(h) nós.(h) (::Z)",
             "end": "",
             "start": "(c4) ",
-            "default": "(hr hr hr) (f) (h) (h) (h) (:)",
+            "default": "(h) (hr hr hr) (f) (h) (h) (h) (:)",
             "patterns": [
                 {
                     "symbol": "*",
-                    "gabc": "(hr hr hr) (fgwhr1) (hr) (h) (hr1) (hr) (h) (::)"
+                    "gabc": "(h) (hr hr hr) (fgwhr1) (hr) (h) (hr1) (hr) (h) (::)"
                 },
                 {
                     "symbol": "?",
@@ -177,7 +177,7 @@
                 },
                 {
                     "symbol": ":",
-                    "gabc": "(hr hr hr) (h) (:)"
+                    "gabc": "(h) (hr hr hr) (h) (:)"
                 }
             ],
             "find": [],
@@ -191,11 +191,11 @@
             "optional_start": "(g) (h)",
             "end": "",
             "start": "(c4) ",
-            "default": "(hr hr hr) (f) (gh) (hr1) (gr) (f) (::) ",
+            "default": "(h) (hr hr hr) (f) (gh) (hr1) (gr) (f) (::) ",
             "patterns": [
                 {
                     "symbol": "*",
-                    "gabc": "(hr hr hr) (g) (hr1) (fr) (f) (;) "
+                    "gabc": "(h) (hr hr hr) (g) (hr1) (fr) (f) (;) "
                 },
                 {
                     "symbol": "&",
@@ -203,7 +203,7 @@
                 },
                 {
                     "symbol": "+",
-                    "gabc": "(hr hr hr) (hr1) (gr) (g)"
+                    "gabc": "(h) (hr hr hr) (hr1) (gr) (g)"
                 }
             ],
             "find": [],
