+++ output.gabc
@@ -1,0 +1,113 @@
+(c3) <sp>V/</sp> O(g)RA(g)ÇÃO(g) EU(g)CA(g)RÍS(g)TI(g)CA(g) II
+
+R.(g) (E(g)le(g) es(g)tá(g) no(g) mei(g)o(g) de(g) nós.)
+V.(g) Co(g)ra(g)ções(g) ao(g) al(g)to.
+R.(g) (O(g) nos(g)so(g) co(g)ra(g)ção(g) es(g)tá(g) em(g) Deus.)
+V.(g) De(g)mos(g) gra(g)ças(g) ao(g) Senh(g)or,(g) nos(g)so(g) Deus.
+R.(g) (É(g) nos(g)so(g) de(g)ver(g) e(g) nos(g)sa(g) sal(g)va(g)ção.)
+(g)Na(g) ver(g)da(g)de,(g) é(g) dig(g)no(g) e(g) jus(g)to,
+(g)é(g) nos(g)so(g) de(g)ver(g) e(g) sal(g)va(g)ção(g) dar-(g)vos(g) gra(g)ças
+(g)sem(g)pre(g) e(g) em(g) to(g)do(g) lu(g)gar,
+(g)Senh(g)or,(g) Pai(g) san(g)to,
+(g)por(g) vos(g)so(g) a(g)ma(g)do(g) Filh(g)o,(g) Je(g)sus(g) Cris(g)to.
+(g)E(g)le(g) é(g) a(g) vos(g)sa(g) Pa(g)la(g)vra,
+(g)pe(g)la(g) qual(g) tu(g)do(g) cri(g)as(g)te(g)s.
+E(g)le(g) é(g) o(g) nos(g)so(g) Sal(g)va(g)dor(g) e(g) Re(g)den(g)to(g)r,
+que(g) se(g) en(g)car(g)nou(g) pe(g)lo(g) Es(g)pí(g)ri(g)to(g) San(g)to
+(g)e(g) nas(g)ceu(g) da(g) Vir(g)gem(g) Ma(g)ri(g)a.
+(g)E(g)le,(g) pa(g)ra(g) cum(g)prir(g) a(g) vos(g)sa(g) von(g)ta(g)de
+(g)e(g) a(g)dqui(g)rir(g) pa(g)ra(g) vós(g) um(g) po(g)vo(g) san(g)to,
+(g)es(g)ten(g)deu(g) os(g) bra(g)ços(g) na(g) ho(g)ra(g) da(g) su(g)a(g) pai(g)xão,
+(g)a(g) fim(g) de(g) ven(g)cer(g) a(g) mor(g)te
+(g)e(g) ma(g)ni(g)fes(g)tar(g) a(g) res(g)sur(g)rei(g)ção.
+(g)Por(g) is(g)so,
+(g)com(g) os(g) An(g)jos(g) e(g) to(g)dos(g) os(g) San(g)tos,
+(g)pro(g)cla(g)ma(g)mos(g) vos(g)sa(g) gló(g)ri(g)a,
+(g)can(g)tan(g)do(g) a(g) u(g)ma(g) só(g) voz:
+
+(g)San(g)to,(g) San(g)to,(g) San(g)to,
+(g)Senh(g)or,(g) Deus(g) do(g) u(g)ni(g)ver(g)so.
+(g)O(g) céu(g) e(g) a(g) ter(g)ra(g) pro(g)cla(g)mam(g) a(g) vos(g)sa(g) gló(g)ri(g)a.
+H(g)o(g)sa(g)na(g) nas(g) al(g)tu(g)ras!
+(g)Ben(g)di(g)to(g) o(g) que(g) vem(g) em(g) no(g)me(g) do(g) Senh(g)o(g)r!
+Ho(g)sa(g)na(g) nas(g) al(g)tu(g)ra(g)s!
+
+<al(g)t>O(g) sa(g)cer(g)do(g)te,(g) de(g) bra(g)ços(g) a(g)ber(g)tos,(g) di(g)z:</alt>
+<(g)c>C(g)P</c>(g) Na(g) ver(g)da(g)de,(g) ó(g) Pai,(g) vós(g) sois(g) San(g)to,
+(g)fon(g)te(g) de(g) to(g)da(g) san(g)ti(g)da(g)de.
+
+<(g)al(g)t>U(g)ne(g) as(g) mãos(g) e,(g) es(g)ten(g)den(g)do-(g)as(g) so(g)bre(g) as(g) o(g)fe(g)ren(g)das,(g) di(g)z:</alt>
+<(g)c>C(g)C</c>(g) San(g)ti(g)fi(g)cai,(g) pois,(g) es(g)tes(g) dons,
+(g)der(g)ra(g)man(g)do(g) so(g)bre(g) e(g)les(g) o(g) vos(g)so(g) Es(g)pí(g)ri(g)to,
+<(g)al(g)t>u(g)ne(g) as(g) mãos(g) e(g) tra(g)ça(g) o(g) si(g)nal(g) da(g) cruz,(g) ao(g) mes(g)mo(g) tem(g)po(g) so(g)bre(g) o(g) pão(g) e(g) o(g) cá(g)li(g)ce,(g) di(g)zen(g)do:</(g)al(g)t>
+a(g) fim(g) de(g) que(g) se(g) tor(g)nem(g) pa(g)ra(g) nó(g)s
+o(g) Cor(g)po(g) e(g) o(g) Sa(g)ngue(g) de(g) nos(g)so(g) Senh(g)or(g) Je(g)sus(g) Cris(g)to.
+<(g)al(g)t>U(g)ne(g) as(g) mão(g)s.</al(g)t>
+<al(g)t>A(g) as(g)sem(g)blei(g)a(g) a(g)cla(g)ma:</(g)al(g)t>
+En(g)vi(g)ai(g) o(g) vos(g)so(g) Es(g)pí(g)ri(g)to(g) San(g)to!
+
+(g)NAR(g)RA(g)TI(g)VA
+
+(g)MIS(g)TE(g)RI(g)O(g) DA(g) FE
+
+
+<(g)al(g)t>O(g) sa(g)cer(g)do(g)te,(g) de(g) bra(g)ços(g) a(g)ber(g)tos,(g) di(g)z:</alt>
+
+<(g)c>C(g)C</c>(g) Ce(g)le(g)bran(g)do,(g) pois,(g) o(g) me(g)mo(g)ri(g)al
+(g)da(g) mor(g)te(g) e(g) res(g)sur(g)rei(g)ção(g) do(g) vos(g)so(g) Filh(g)o,
+(g)nós(g) vos(g) o(g)fe(g)re(g)ce(g)mos,(g) ó(g) Pai,
+(g)o(g) Pão(g) da(g) vi(g)da(g) e(g) o(g) Cá(g)li(g)ce(g) da(g) sal(g)va(g)ção;
+(g)e(g) vos(g) a(g)gra(g)de(g)ce(g)mos
+(g)po(g)rque(g) nos(g) tor(g)nas(g)tes(g) dig(g)nos
+(g)de(g) es(g)tar(g) aqui(g) na(g) vos(g)sa(g) pre(g)sen(g)ça(g) e(g) vos(g) ser(g)vi(g)r.
+<al(g)t>A(g) as(g)sem(g)blei(g)a(g) a(g)cla(g)ma:</(g)al(g)t>
+
+A(g)cei(g)tai,(g) ó(g) Senh(g)or,(g) a(g) nos(g)sa(g) o(g)fer(g)ta!
+
+(g)Su(g)pli(g)can(g)tes,(g) vos(g) pe(g)di(g)mo(g)s
+que,(g) par(g)ti(g)ci(g)pan(g)do(g) do(g) Cor(g)po(g) e(g) Sa(g)ngue(g) de(g) Cris(g)to,
+(g)se(g)ja(g)mos(g) reu(g)ni(g)dos(g) pe(g)lo(g) Es(g)pí(g)ri(g)to(g) San(g)to
+(g)num(g) só(g) cor(g)po.
+
+<(g)al(g)t>A(g) as(g)sem(g)blei(g)a(g) a(g)cla(g)ma:</(g)al(g)t>
+
+O(g) Es(g)pí(g)ri(g)to(g) nos(g) u(g)na(g) num(g) só(g) cor(g)po!
+
+<c>1(g)C</c>(g) Lem(g)brai-(g)vos,(g) ó(g) Pai,
+(g)da(g) vos(g)sa(g) I(g)gre(g)ja(g) que(g) se(g) faz(g) pre(g)sen(g)te(g) pe(g)lo(g) mun(g)do(g) in(g)tei(g)ro;*(g) que(g) e(g)la(g) cres(g)ça(g) na(g) ca(g)ri(g)da(g)de,
+(g)em(g) co(g)munh(g)ão(g) com(g) o(g) Pa(g)pa(g) N.,
+(g)com(g) o(g) nos(g)so(g) Bis(g)po(g) N.,*
+os(g) bis(g)pos(g) do(g) mun(g)do(g) in(g)tei(g)ro,
+(g)os(g) pres(g)bí(g)te(g)ros,(g) os(g) di(g)á(g)co(g)no(g)s
+e(g) to(g)dos(g) os(g) mi(g)nis(g)tros(g) do(g) vos(g)so(g) po(g)vo.
+<(g)al(g)t>A(g) as(g)sem(g)blei(g)a(g) a(g)cla(g)ma:</(g)al(g)t>
+
+Lem(g)brai-(g)vos,(g) ó(g) Pai,(g) da(g) vos(g)sa(g) I(g)gre(g)ja!
+
+
+<c>2(g)C</c>(g) Lem(g)brai-(g)vos(g) tam(g)bém,(g) na(g) vos(g)sa(g) mi(g)se(g)ri(g)cór(g)di(g)a,
+(g)dos(g) nos(g)sos(g) ir(g)mãos(g) e(g) ir(g)mã(g)s
+que(g) a(g)dor(g)me(g)ce(g)ram(g) na(g) es(g)pe(g)ran(g)ça(g) da(g) res(g)sur(g)rei(g)ção
+(g)e(g) de(g) to(g)dos(g) os(g) que(g) par(g)ti(g)ram(g) des(g)ta(g) vi(g)da;
+(g)a(g)colh(g)ei-(g)os(g) jun(g)to(g) a(g) vós
+(g)na(g) luz(g) da(g) vos(g)sa(g) fa(g)ce.
+<(g)al(g)t>A(g) as(g)sem(g)blei(g)a(g) a(g)cla(g)ma:</(g)alt>
+
+(g)Con(g)ce(g)dei-(g)lh(g)es,(g) ó(g) Senh(g)or,(g) a(g) luz(g) e(g)ter(g)na!
+
+<c>3(g)C</c>(g) En(g)fim,(g) nós(g) vos(g) pe(g)di(g)mos,
+(g)ten(g)de(g) pi(g)e(g)da(g)de(g) de(g) to(g)dos(g) nó(g)s
+e(g) dai-(g)nos(g) par(g)ti(g)ci(g)par(g) da(g) vi(g)da(g) e(g)ter(g)na,
+(g)com(g) a(g) Vir(g)gem(g) Ma(g)ri(g)a,(g) Mãe(g) de(g) Deus,
+(g)São(g) Jo(g)sé,(g) seu(g) es(g)po(g)so,(g) os(g) A(g)pós(g)to(g)lo(g)s,
+e(g) to(g)dos(g) os(g) San(g)tos(g) que(g) nes(g)te(g) mun(g)do
+(g)vi(g)ve(g)ram(g) na(g) vos(g)sa(g) a(g)mi(g)za(g)de,
+(g)a(g) fim(g) de(g) vos(g) lou(g)var(g)mos(g) e(g) glo(g)ri(g)fi(g)car(g)mo(g)s
+<al(g)t>u(g)ne(g) as(g) mão(g)s</alt>
+
+(g)por(g) Je(g)sus(g) Cris(g)to,(g) vos(g)so(g) Filh(g)o.
+
+<(g)al(g)t>E(g)rgue(g) a(g) pa(g)te(g)na(g) com(g) a(g) hós(g)ti(g)a(g) e(g) o(g) cá(g)li(g)ce,(g) di(g)zen(g)do:</(fe)alt>
+
+
+(ef)POR(g) CRIS(fg)TO(f) (:)
